Warning fix

Make "all warnings are errors" be the default behavior in the library

Repair ALL warnings.

See merge request !209

+++ src/core/demo/SamplingSource.cpp
@@ -1,5 +1,5 @@
 #include "math/discretegaussiangenerator.h"
-#include "utils/debug.h";
+#include "utils/debug.h"
 //#include <vld.h>
 using namespace lbcrypto;
 
@@ -16,14 +16,14 @@ int main() {
 	std::cout << "Probability matrix generation: " << finish - start << " ms\n";
 
 	start = currentDateTime();
-	for (int i = 0;i < count;i++) {
+	for (size_t i = 0;i < count;i++) {
 		dgg.GenerateInteger(0, std, 1024);
 	}
 	finish = currentDateTime();
 	std::cout << "Sampling 100000 integers (Rejection): " << finish - start << " ms\n";
 
 	start = currentDateTime();
-	for (int i = 0;i < count;i++) {
+	for (size_t i = 0;i < count;i++) {
 		//dgg.GenerateIntegerKnuthYao();
 	}
 	finish = currentDateTime();
